DataExplorer/requirements.txt
Django>=1.10.0,<1.10.99
python-quickbooks
